BrowserRouter,Provider kurulumlari
