+++ package.json
@@ -6,6 +6,7 @@
     "react": "^16.8.3",
     "react-dom": "^16.8.3",
     "react-redux": "^6.0.1",
+    "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.5",
     "react-thunk": "^1.0.0",
     "redux": "^4.0.1",
