+++ src/Styles/App.css
@@ -1,16 +1,16 @@
 .App {
-  text-align: center;
+  text-align: left;
 }
 
 .App-logo {
-  animation: App-logo-spin infinite 20s linear;
-  height: 40vmin;
+  /* animation: App-logo-spin infinite 20s linear; */
+  height: 20vmin;
   pointer-events: none;
 }
 
 .App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+  background-color:darkgray;
+  min-height: 10vh;
   display: flex;
   flex-direction: column;
   align-items: center;
