+++ src/components/pages/AccountsPage.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+class AccountsPage extends Component {
+    render() {
+        return (
+            <div>
+                <h2>Accounts List</h2>
+            </div>
+        );
+    }
+} 
+
+export default AccountsPage;
