+++ src/index.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import './Styles/index.css';
+import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import {createStore,applyMiddleware} from 'redux';
 import rootReducer from './reducers/rootReducer';
 import thunk from 'redux-thunk';
 import {composeWithDevTools} from 'redux-devtools-extension';
-
+import {Provider} from 'react-redux';
+import {BrowserRouter} from 'react-router-dom';
 
 const store = createStore(
   rootReducer,
@@ -18,7 +19,11 @@ const store = createStore(
 
 ReactDOM.render 
 ( 
-    <App />, 
+    <BrowserRouter>
+    <Provider store={store}>
+      <App />
+    </Provider>
+    </BrowserRouter>, 
     document.getElementById('root')
 );
